Organization Backend Functionality
Updated Modules of Create Organization, Update Organization by the one who created it, Delete organization by the one who created it and Get all Organizations.

+++ server/controllers/organizationControllers/deleteOrganization.js
@@ -1,4 +1,5 @@
 import db from "../../config/db.js";
+import jwt from 'jsonwebtoken';
 
 export const deleteOrganization  = async (req, res) => {
 
@@ -6,9 +7,74 @@ export const deleteOrganization  = async (req, res) => {
         
         const {id} = req.params;
 
+        if(!id){
+            return res.status(400).send({
+                success: false,
+                message: "Organization ID required"
+            });
+        }
+
+        // Check if the Authorization header exists
+        let token = req.header("Authorization");
+        if (!token) {
+            return res.status(401).send({ message: "Authorization token is required" });
+        }
+        
+        // Verify the token and extract the user ID
+        const decoded = jwt.verify(token, process.env.JWT_SECRET);
+        const userId = decoded.id;
+        
+        if (!userId) {
+            return res.status(401).send({ message: "Invalid token" });
+        }
+
+
+        // Check if organization exist
+        const checkOrganizationExist = `
+            SELECT * FROM organizations WHERE organization_id = ?;
+        `;
+        const organizationExist = await new Promise((resolve, reject) => {
+            db.query(checkOrganizationExist, [id], (err, results) => {
+                if (err) {
+                    reject(err);
+                } else {
+                    resolve(results[0]);
+                }
+            });
+        });
+
+        if (!organizationExist) {
+            return res.status(404).send({
+                success: false,
+                message: "There is no such Organization Exist in the Database",
+            });
+        }
+
+
+        // Check if the organization is created by the user
+        const checkOrganizationCreatedUser = `
+            SELECT * FROM organizations WHERE organization_id = ? AND created_by = ?;
+        `;
+        const org = await new Promise((resolve, reject) => {
+            db.query(checkOrganizationCreatedUser, [id, userId], (err, results) => {
+                if (err) {
+                    reject(err);
+                } else {
+                    resolve(results[0]);
+                }
+            });
+        });
+
+        if (!org) {
+            return res.status(404).send({
+                success: false,
+                message: "You do not have permission to delete this Organization",
+            });
+        }
+
         const deleteOrganizationQuery = `DELETE FROM organizations WHERE organization_id = ?;`;
 
-        const result = await new Promise((resolve, reject) => {
+        await new Promise((resolve, reject) => {
             db.query(deleteOrganizationQuery, [id], (err, results) => {
                 if(err){
                     reject(err);
@@ -18,13 +84,6 @@ export const deleteOrganization  = async (req, res) => {
                 }
             });
         });
-
-        if(result.affectedRows === 0){
-            return res.status(400).send({
-                success: false,
-                message: "Data cannot be Deleted as Organization with this ID does not exist"
-            });
-        }
 
         return res.status(200).send({
             success: true,
